+++ src/main/java/kitchenpos/ui/dto/menu/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.menu;
 
 import java.math.BigDecimal;
 import java.util.List;
