+++ src/main/java/kitchenpos/ui/dto/menu/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.menu;
 
 public class MenuGroupCreateRequest {
