+++ src/main/java/kitchenpos/ui/dto/ordertable/OrderTableIdDto.java
@@ -1,0 +1,18 @@
+package kitchenpos.ui.dto.ordertable;
+
+public class OrderTableIdDto {
+
+    private Long id;
+
+    OrderTableIdDto() {
+
+    }
+
+    public OrderTableIdDto(final Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
