+++ src/main/java/kitchenpos/ui/dto/tablegroup/TableGroupUngroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.tablegroup;
 
 public class TableGroupUngroupRequest {
