+++ src/test/java/kitchenpos/support/fixture/dto/MenuGroupCreateRequestFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.support.fixture.dto;
 
-import kitchenpos.ui.dto.MenuGroupCreateRequest;
+import kitchenpos.ui.dto.menu.MenuGroupCreateRequest;
 
 public class MenuGroupCreateRequestFixture {
